feat: 抽出 ASRService 管理 worker 生命周期
新增 asr_service.py，集中 ASR worker 配置构造、缺失模型判断、下载后校验、加载切换和失败恢复逻辑。

LiveTranslateApp 不再直接持有 ASRClient，仅负责触发下载/加载对话框、重置增量 ASR 状态和更新 overlay 状态显示。

转写、语言切换、padding 更新、ready 状态和 shutdown 统一通过 ASRService 公开接口访问，为后续 PipelineController 拆分降低 main.py 的运行时耦合。
